This commit moves the `read.c` gruntwork under a `struct io` interface. This groups I/O-layer operations together, providing a module-style interface (perhaps to be followed by writing, in the future).
`cl_accept()` is now passed an fd, which is passed along to the I/O layer for the sake of termios control and the like.

Paul Evan's LibTermKey has been brought in for `ecma48.c`. I worked with him to add a `termkey_push_bytes()` interface to his library, along with a few internal changes. This permits sending in data from an arbitary source, as we have in the case of libcl. I'm developing against his bzr trunk, but `termkey_push_bytes()` will make it into 0.12, when he releases that.

I've made a start on mapping out some UI events, populated by TermKeyKey inputs.

The `read.c` state is now private to `read.c`, at the cost of a little extra work for allocation.