Two problems here:
Firstly, resizing needs to occur if any of the incoming bytes overflow the current buffer, not just if the first byte does. I knew I'd get this wrong. So previously this was `if ((*count + n) % blocksz == 0)` and now we reallocate `if ((*count - 1) % blocksz < n)` which I'm intending to be literally "if the available space is less than the incoming length".

Secondly `*count` may live inside the reallocated space, and so we need to avoid dereferencing it after reallocation. This is pretty ugly, but I still think this routine is worth keeping generalised to both situations where it's used.