Rework of the I/O layer. This commit implements calling through to the next handler to form "chains" which represent wrapping of one protocl inside another.
I/O chains are traversed in one direction for reading, and in the opposite direction for writing. Each handler may maintain its own `ioctx` state, holding whatever private state it needs.

The idea is that each I/O handler can interject at any point along a chain and handle what it needs to, or has the option to delegate by passing its call along to the next handler. This might be best explained by a diagram.

This is all humongously complex. I've attempted to at least keep it self-contained. I've done my best to make it as clear as I can. Perhaps it'll be clearer with more protocols implemented, as a few things are unneccessary for the degenerative cases.

I've removed `OUT_PRINTF` (and therfore `cl_output` is now just an enum), and split the printf style functions off seperately, as I think this is clearer.