Introduced telnet. Added `.motd()` and split off post-`cl_accept()` creation to a new interface, `cl_ready()`.
This sees a signifiant change to the I/O layer's creation and destruction; `.create()` and `.destroy()` now chain together just as the I/O functions themselves do. This provides a mechanism for postponing creation of a specific I/O handler until its dependencies are met.

The motivating case for this change is that the ecma48 layer depends on telnet having negotiated an appropriate `$TERM` type.

This commit brings in a dependency on `libtelnet`. Currently this is a fork I made on github, which provies printf-like interfaces with `va_list` equivalents to their variadic counterparts. At the time of writing, the pull request for that patch hasn't made it into Sean Middleditch's master branch for libtelnet, yet.